+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix: Wifi::disconnect acts on connections instead of interface
 - Fix: In connection name command usages, guard in ticks if it contains spacings
 - Fix: Wifi::set_metric would use connection name instead of ssid
+- Feat: Add Wifi::auto_connect
 
 # 0.3.2 - 06/05/2025
 - Fix: Wifi::up